requirements.txt
elasticsearch>=8.0,<9.0
sentence-transformers
openai
openpyxl
python-docx
pdfplumber
tiktoken
python-dotenv
